chore(react-motions): rename factories
