+++ change/@fluentui-react-motions-preview-18594f5a-3076-4d8a-843c-c8ce195de830.json
@@ -1,0 +1,7 @@
+{
+  "type": "minor",
+  "comment": "BREAKING: rename createAtom() to createMotionComponent() & createPresence() to createPresenceMotion()",
+  "packageName": "@fluentui/react-motions-preview",
+  "email": "olfedias@microsoft.com",
+  "dependentChangeType": "patch"
+}
