+++ packages/react-components/react-motions-preview/etc/react-motions-preview.api.md
@@ -15,10 +15,10 @@ export type AtomMotion = {
 export type AtomMotionFn = (element: HTMLElement) => AtomMotion;
 
 // @public
-export function createAtom(motion: AtomMotion | AtomMotionFn): React_2.FC<AtomProps>;
+export function createMotionComponent(motion: AtomMotion | AtomMotionFn): React_2.FC<AtomProps>;
 
 // @public (undocumented)
-export function createPresence(motion: PresenceMotion | PresenceMotionFn): React_2.FC<PresenceProps>;
+export function createPresenceComponent(motion: PresenceMotion | PresenceMotionFn): React_2.FC<PresenceProps>;
 
 // @public (undocumented)
 const durationFast = 150;
