--- packages/react-components/react-motions-preview/stories/Motion/CreateAtom.stories.md
@@ -1,1 +1,0 @@
-`createAtom()` is a factory function that creates a React component based on the provided atom definition. This component can be used to animate any element.
