+++ packages/react-components/react-motions-preview/stories/Motion/CreateMotionComponent.stories.tsx
@@ -1,9 +1,9 @@
 import { makeStyles, shorthands, tokens, Label, Slider, useId } from '@fluentui/react-components';
-import { createAtom, motionTokens } from '@fluentui/react-motions-preview';
+import { createMotionComponent, motionTokens } from '@fluentui/react-motions-preview';
 import type { MotionImperativeRef } from '@fluentui/react-motions-preview';
 import * as React from 'react';
 
-import description from './CreateAtom.stories.md';
+import description from './CreateMotionComponent.stories.md';
 
 const useClasses = makeStyles({
   container: {
@@ -53,16 +53,16 @@ const useClasses = makeStyles({
   },
 });
 
-const FadeEnter = createAtom({
+const FadeEnter = createMotionComponent({
   keyframes: [{ opacity: 0 }, { opacity: 1 }],
   duration: motionTokens.durationSlow,
 });
-const FadeExit = createAtom({
+const FadeExit = createMotionComponent({
   keyframes: [{ opacity: 1 }, { opacity: 0 }],
   duration: motionTokens.durationSlow,
 });
 
-export const CreateAtom = () => {
+export const CreateMotionComponent = () => {
   const classes = useClasses();
   const sliderId = useId();
 
@@ -117,7 +117,7 @@ export const CreateAtom = () => {
   );
 };
 
-CreateAtom.parameters = {
+CreateMotionComponent.parameters = {
   docs: {
     description: {
       story: description,
