+++ packages/react-components/react-motions-preview/stories/Motion/CreatePresenceComponent.stories.md
@@ -1,0 +1,1 @@
+`createPresenceComponent()` is a factory function that creates a React component based on the provided presence definition. This component can be used to animate any element and intended to have a state via the `visible` prop.
