+++ packages/react-components/react-motions-preview/stories/Motion/CustomMotion.stories.md
@@ -1,4 +1,4 @@
-You can create completely custom motions, this is applicable to both `createAtom()` and `createPresence()` factories.
+You can create completely custom motions, this is applicable to both `createMotionComponent()` and `createPresenceComponent()` factories.
 
 Atoms are defined by `AtomMotion` interface and presence animations are defined by `PresenceMotion` interface.
 `AtomMotion` contains definitions for `keyframes` and its options that are used internally to create [`KeyframeEffect`](https://developer.mozilla.org/en-US/docs/Web/API/KeyframeEffect) for an animation.
