+++ packages/react-components/react-motions-preview/stories/Motion/ImperativeRefPlayState.stories.md
@@ -1,3 +1,3 @@
 By default, the child component will be animated when it first mounts. The state of a motion can be controlled using `setPlayState()` via `imperativeRef` prop.
 
-`imperativeRef` works with both `createAtom()` and `createPresence()` factories.
+`imperativeRef` works with both `createMotionComponent()` and `createPresenceComponent()` factories.
