+++ packages/react-components/react-motions-preview/stories/Motion/MotionDefault.stories.tsx
@@ -1,5 +1,5 @@
 import { makeStyles, shorthands, tokens } from '@fluentui/react-components';
-import { createAtom, motionTokens } from '@fluentui/react-motions-preview';
+import { createMotionComponent, motionTokens } from '@fluentui/react-motions-preview';
 import type { MotionImperativeRef } from '@fluentui/react-motions-preview';
 import * as React from 'react';
 
@@ -23,7 +23,7 @@ const useClasses = makeStyles({
   },
 });
 
-const FadeEnter = createAtom({
+const FadeEnter = createMotionComponent({
   keyframes: [{ opacity: 0 }, { opacity: 1 }],
   duration: motionTokens.durationSlow,
 });
