+++ packages/react-components/react-motions-preview/stories/Motion/MotionFunctions.stories.md
@@ -1,7 +1,7 @@
 Atoms and presence definitions can be also defined as functions that accept an animated element as an argument. This allows to define more complex animations that depend on the animated element's properties, for example:
 
 ```ts
-const Grow = createAtom(element => ({
+const Grow = createMotionComponent(element => ({
   duration: 300,
   keyframes: [
     { opacity: 0, maxHeight: `${element.scrollHeight / 2}px` },
