+++ packages/react-components/react-motions-preview/stories/Motion/TokensUsage.stories.tsx
@@ -1,5 +1,5 @@
 import { makeStyles, shorthands, tokens } from '@fluentui/react-components';
-import { createAtom } from '@fluentui/react-motions-preview';
+import { createMotionComponent } from '@fluentui/react-motions-preview';
 import * as React from 'react';
 
 import description from './TokensUsage.stories.md';
@@ -31,7 +31,7 @@ const useClasses = makeStyles({
   },
 });
 
-const BackgroundChange = createAtom({
+const BackgroundChange = createMotionComponent({
   keyframes: [
     { backgroundColor: tokens.colorStatusDangerBackground3 },
     { backgroundColor: tokens.colorStatusSuccessBackground3 },
