+++ packages/react-components/react-motions-preview/stories/Motion/index.stories.ts
@@ -2,15 +2,15 @@ import MotionDescription from './MotionDescription.md';
 
 export { MotionDefault } from './MotionDefault.stories';
 
-export { CreateAtom as createAtom } from './CreateAtom.stories';
-export { CreatePresence as createPresence } from './CreatePresence.stories';
+export { CreateMotionComponent as createMotionComponent } from './CreateMotionComponent.stories';
+export { CreatePresenceComponent as createPresenceComponent } from './CreatePresenceComponent.stories';
 
 export { PresenceAppear as appear } from './PresenceAppear.stories';
 export { PresenceUnmountOnExit as unmountOnExit } from './PresenceUnmountOnExit.stories';
 
 export { ImperativeRefPlayState as setPlayState } from './ImperativeRefPlayState.stories';
 
-export { CustomAtom } from './CustomAtom.stories';
+export { CustomMotion } from './CustomMotion.stories';
 export { TokensUsage } from './TokensUsage.stories';
 export { MotionFunctions } from './MotionFunctions.stories';
